make install.py call pip install .
